agent/requirements.txt
psutil==5.9.8
pyyaml==6.0.1
requests==2.31.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0

# Testing
pytest==7.4.4
pytest-mock==3.12.0
